GET INT NEEDS TO BE SWITCHED TO GET LONG FOR THE UNIX CONVERTER TO WORK

+++ src/main/java/data_access/RiotUserDataAccessObject.java
@@ -69,7 +69,7 @@ public class RiotUserDataAccessObject {
                     playerData.getBoolean("win"),
                     gameInfo.getInt("gameDuration"),
                     gameInfo.getString("gameMode"),
-                    gameInfo.getInt("gameCreation")));
+                    gameInfo.getLong("gameCreation")));
         }
         return matchHistoryFactory.createMatchHistory(matchList);
     }
